refactor: NaverMap 컴포넌트 명칭 수정 - `DynamicMap`으로 수정
